+++ src/rex.forms/src/rex/forms/interface/__init__.py
@@ -3,22 +3,12 @@
 #
 
 
-from .channel import *
-from .completion_processor import *
 from .draftform import *
-from .entry import *
 from .form import *
-from .parameter_supplier import *
-from .task import *
 
 
 __all__ = (
-    'Channel',
-    'Form',
-    'Task',
-    'Entry',
-    'ParameterSupplier',
-    'TaskCompletionProcessor',
     'DraftForm',
+    'Form',
 )